moving chainport types to new file
