+++ ironfish-cli/src/utils/chainport/index.ts
@@ -4,3 +4,4 @@
 
 export * from './chainport'
 export * from './metadata'
+export * from './types'
