+++ ironfish-cli/src/utils/chainport/types.ts
@@ -1,0 +1,68 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+/**
+ * This file contains the types for the response from the chainport API
+ */
+
+export type ChainportBridgeTransaction = {
+  bridge_output: {
+    publicAddress: string
+    amount: string
+    memoHex: string
+    assetId: string
+  }
+  gas_fee_output: {
+    publicAddress: string
+    amount: string
+    memo: string
+  }
+  bridge_fee: {
+    source_token_fee_amount: number | null
+    portx_fee_amount: number
+    is_portx_fee_payment: boolean
+  }
+}
+
+export type ChainportNetwork = {
+  chainport_network_id: number
+  shortname: string
+  name: string
+  chain_id: number
+  explorer_url: string
+  label: string
+  blockchain_type: string
+  native_token_symbol: string
+  network_icon: string
+}
+
+export type ChainportVerifiedToken = {
+  decimals: number
+  id: number
+  name: string
+  pinned: boolean
+  web3_address: string
+  symbol: string
+  token_image: string
+  target_networks: number[]
+  chain_id: number | null
+  network_name: string
+  network_id: number
+  blockchain_type: string
+  is_stable: boolean
+  is_lifi: boolean
+}
+
+export type ChainportTransactionStatus = {
+  base_network_id?: number
+  base_tx_hash?: string
+  base_tx_status?: number
+  base_token_address?: string
+  target_network_id?: number
+  target_tx_hash?: string
+  target_tx_status?: number
+  target_token_address?: string
+  created_at?: string
+  port_in_ack?: boolean
+}
